changes the task required argument from module to action (#1)
This commit changes the task argument `module` to `action` to better
align with the Task() structure.

This change fixes a bug where the previous task args would be carried
over to a new task in the current play.  When the new task is created,
it should not have any args by default.