refactors attrs and updates models
This commit is an intrusive one that refactors the attrs module to
simplify and standardize the code path.  Attributes are now broken out
into seperate class objects insteading of using a single Attribute
class.  All attribute value validation has been moved to validators.
This change will make the code more readable and deterministic.

This PR has a dependency on merging #31

Signed-off-by: Peter Sprygada <psprygad@redhat.com>