adds new role bindings feature to network-runner
This commit adds a new feature to network-runner that will allow Python
objects to be dynamically built based on a bindings file in the role.
The bindings file is responsible for generating the exposed methods and
signatures for the dynamic object.  Each method will return a Task
instance that can be directly inserted into a Play object.

Signed-off-by: Peter Sprygada <psprygad@redhat.com>