Remove the use of "default interface" on EOS
Using "default interface" to clear configurations on each port
causes issues on EOS when the port must have a speed set in the interface
stanza due to not conforming to the switch's default breakout mode. For example,
configuring a port that is set to 40G will revert it back to 4x10G mode on a
7050QX switch. Simply adding a speed option would cause the entire switch
to stop forwarding when any port is moved from breakout to non-breakout mode,
so this patch removes the use of "default interface" in favor of explicitly clearing
relevant configurations.